requirements.txt
streamlit==1.32.0
fastapi==0.110.0
uvicorn==0.27.1
transformers==4.38.2
torch==2.2.1
pandas==2.2.1
plotly==5.19.0
sqlalchemy==2.0.27
python-dotenv==1.0.1
requests==2.31.0
mistralai==0.0.12 